+++ crosshatch/KnownAssets/USDT0.ts
@@ -1,4 +1,5 @@
-import { EvmAdapter, EvmAsset } from "../Evm/Evm.ts"
+import { EvmAsset } from "../Evm/Evm.ts"
+import { Eip3009Adapter } from "../payload_adapters/Erc3009.ts"
 import { type PhysicalAsset } from "../PhysicalAsset.ts"
 
 export const USDT0 = {
@@ -11,14 +12,14 @@ export const USDT0 = {
         decimals: 6,
         name: "USDT0",
         version: "1",
-        adapter: EvmAdapter.EvmAdapter,
+        adapters: [Eip3009Adapter],
       },
       2201: {
         asset: EvmAsset.EvmAsset.make("0x78Cf24370174180738C5B8E352B6D14c83a6c9A9"),
         decimals: 6,
         name: "USDT0",
         version: "1",
-        adapter: EvmAdapter.EvmAdapter,
+        adapters: [Eip3009Adapter],
       },
     },
   },
